Issue #142 refactor: 일주일 스케줄 설정하기 viewModel로 이동
